// frontend/src/lib/api/model/pythagoreItem.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Explorito - Application Éducative
 * OpenAPI spec version: 1.0.0
 */

/**
 * Une question répondue : opérandes + réponse donnée par l'enfant.
 */
export interface PythagoreItem {
  /**
   * Premier opérande
   * @minimum 1
   * @maximum 12
   */
  a: number;
  /**
   * Second opérande
   * @minimum 1
   * @maximum 12
   */
  b: number;
  /** Réponse donnée par l'enfant */
  answer: number;
}